# 2x Faster `bf16` Euclidean Distance on AMD Genoa